requirements.txt
kafka-python
Flask
twilio
requests
ddgs
psycopg2-binary
